requirements.txt
torch
torchvision
torchaudio
transformers
accelerate
datasets
opencv-python
lerobot
pytorch-lightning
wandb
faiss-cpu
scikit-learn
einops
tqdm
numpy
Pillow
pyyaml

